+++ tt-media-server/cpp_server/src/api/llm_controller.cpp
@@ -271,7 +271,7 @@ LLMController::makeStreamingCallback(std::shared_ptr<ResponseWriter> writer,
 
     // Feed token to prefix accumulator for incremental index updates
     if (session && !chunk.choices.empty() && chunk.choices[0].token_id) {
-      session->addGeneratedToken(*chunk.choices[0].token_id);
+      session->addGeneratedToken(static_cast<int>(*chunk.choices[0].token_id));
     }
 
     if (!chunk.choices.empty()) writer->handleTokenChunk(chunk);
