+++ tt-media-server/cpp_server/src/utils/block_hash_accumulator.cpp
@@ -12,7 +12,7 @@ namespace tt::utils {
 
 BlockHashAccumulator::BlockHashAccumulator(
     std::vector<uint64_t> initialHashes,
-    std::vector<int64_t> partialBlockTokens)
+    std::vector<int> partialBlockTokens)
     : hashes_(std::move(initialHashes)),
       blockBuffer_(std::move(partialBlockTokens)),
       firstBlockSize_(config::kvCacheFirstBlockSize()),
@@ -21,7 +21,7 @@ BlockHashAccumulator::BlockHashAccumulator(
 }
 
 std::optional<std::vector<uint64_t>> BlockHashAccumulator::addToken(
-    int64_t tokenId) {
+    int tokenId) {
   blockBuffer_.push_back(tokenId);
 
   size_t targetSize = hashes_.empty() ? firstBlockSize_ : blockSize_;
@@ -30,8 +30,9 @@ std::optional<std::vector<uint64_t>> BlockHashAccumulator::addToken(
   }
 
   // Block complete - compute hash using chained seeding
+  // Must use sizeof(int) to match conversation_hasher.cpp hashing
   uint64_t newHash = XXH64(blockBuffer_.data(),
-                           blockBuffer_.size() * sizeof(int64_t), parentHash_);
+                           blockBuffer_.size() * sizeof(int), parentHash_);
   hashes_.push_back(newHash);
   parentHash_ = newHash;
   blockBuffer_.clear();
